Improve delete tasks and reuse tls tasks.
